+++ src/components/AdventureManager/__tests__/AdventureManager.list.test.tsx
@@ -1,0 +1,81 @@
+import { screen } from "@testing-library/react";
+import { describe, it, expect, beforeEach, vi } from "vitest";
+import { AdventureManager } from "../AdventureManager";
+import { render } from "@/__tests__/testUtils";
+import * as adventureDatabase from "@/data/adventureDatabase";
+import * as importYaml from "@/utils/importYaml";
+import type { StoredAdventure } from "@/data/adventureDatabase";
+
+// Mock react-router-dom
+const mockNavigate = vi.fn();
+vi.mock("react-router-dom", async () => {
+  const actual = await vi.importActual("react-router-dom");
+  return {
+    ...actual,
+    useNavigate: () => mockNavigate,
+  };
+});
+
+// Mock adventureDatabase
+vi.mock("../../../data/adventureDatabase", async () => {
+  const actual = await vi.importActual("../../../data/adventureDatabase");
+  return {
+    ...actual,
+    listStories: vi.fn(),
+    deleteAdventure: vi.fn(),
+    createAdventure: vi.fn(),
+  };
+});
+
+// Mock importYaml
+vi.mock("../../../utils/importYaml", () => ({
+  importYamlFile: vi.fn(),
+}));
+
+describe("AdventureManager Component", () => {
+  const mockStories: StoredAdventure[] = [
+    {
+      id: "adventure-1",
+      title: "Adventure One",
+      content: "mock content",
+      lastEdited: new Date("2025-11-10T12:00:00"),
+      createdAt: new Date("2025-11-01T12:00:00"),
+    },
+    {
+      id: "adventure-2",
+      title: "Adventure Two",
+      content: "mock content 2",
+      lastEdited: new Date("2025-11-09T12:00:00"),
+      createdAt: new Date("2025-11-02T12:00:00"),
+    },
+  ];
+
+  beforeEach(() => {
+    mockNavigate.mockClear();
+    vi.mocked(adventureDatabase.listStories).mockClear();
+    vi.mocked(adventureDatabase.deleteAdventure).mockClear();
+    vi.mocked(adventureDatabase.createAdventure).mockClear();
+    vi.mocked(importYaml.importYamlFile).mockClear();
+  });
+
+  describe("Adventure List Display", () => {
+    beforeEach(() => {
+      vi.mocked(adventureDatabase.listStories).mockResolvedValue(mockStories);
+    });
+
+    it("renders the new adventure card", async () => {
+      render(<AdventureManager />);
+
+      expect(
+        await screen.findByText("Create a new adventure")
+      ).toBeInTheDocument();
+    });
+
+    it("renders all stories from the database", async () => {
+      render(<AdventureManager />);
+
+      expect(await screen.findByText("Adventure One")).toBeInTheDocument();
+      expect(screen.getByText("Adventure Two")).toBeInTheDocument();
+    });
+  });
+});
