+++ src/components/AdventureTopBar/__tests__/AdventureTopBar.test.tsx
@@ -1,10 +1,4 @@
-import {
-  screen,
-  render,
-  fireEvent,
-  waitFor,
-  act,
-} from "@testing-library/react";
+import { screen, render, fireEvent } from "@testing-library/react";
 import { describe, it, expect, beforeEach, vi } from "vitest";
 import { MemoryRouter } from "react-router-dom";
 import { AdventureTopBar } from "../AdventureTopBar";
@@ -17,6 +11,8 @@ const TEST_STORY_ID = "test-adventure-id";
 
 describe("AdventureTopBar Component", () => {
   beforeEach(async () => {
+    vi.clearAllMocks();
+    vi.restoreAllMocks();
     await setupTestAdventure(TEST_STORY_ID);
   });
 
@@ -135,207 +131,6 @@ describe("AdventureTopBar Component", () => {
 
       const savingIndicator = screen.queryByTestId("saving-indicator");
       expect(savingIndicator).not.toBeInTheDocument();
-    });
-  });
-
-  describe("YAML Download", () => {
-    it("successfully downloads YAML file", async () => {
-      const { mockAdventure } = await import("@/__tests__/mockAdventureData");
-
-      // Mock getAdventure to return valid adventure data
-      const mockGetAdventure = vi
-        .spyOn(await import("@/data/adventureDatabase"), "getAdventure")
-        .mockResolvedValueOnce({
-          id: TEST_STORY_ID,
-          title: mockAdventure.metadata.title,
-          content: "yaml content",
-          lastEdited: new Date(),
-          createdAt: new Date(),
-        });
-
-      // Mock sanitizeFilename
-      const mockSanitizeFilename = vi
-        .spyOn(await import("@/utils/fileDownload"), "sanitizeFilename")
-        .mockReturnValueOnce("mock-test-adventure");
-
-      // Mock downloadFile
-      const mockDownloadFile = vi
-        .spyOn(await import("@/utils/fileDownload"), "downloadFile")
-        .mockImplementationOnce(() => {});
-
-      renderWithAdventure(<AdventureTopBar />, {
-        adventureId: TEST_STORY_ID,
-      });
-
-      const menuButton = await screen.findByTestId(
-        "adventure-context-menu-button"
-      );
-      fireEvent.click(menuButton);
-
-      const downloadMenuItem = screen.getByText("Download adventure as YAML");
-
-      await act(async () => {
-        fireEvent.click(downloadMenuItem);
-      });
-
-      await waitFor(() => {
-        expect(mockGetAdventure).toHaveBeenCalledWith(TEST_STORY_ID);
-      });
-
-      await waitFor(() => {
-        expect(mockSanitizeFilename).toHaveBeenCalledWith(
-          mockAdventure.metadata.title
-        );
-      });
-
-      await waitFor(() => {
-        expect(mockDownloadFile).toHaveBeenCalledWith(
-          "yaml content",
-          "mock-test-adventure.yaml",
-          "text/yaml;charset=utf-8"
-        );
-      });
-
-      mockGetAdventure.mockRestore();
-      mockSanitizeFilename.mockRestore();
-      mockDownloadFile.mockRestore();
-    });
-
-    it("handles early return when adventure is null but adventureId exists", async () => {
-      const { mockAdventure } = await import("@/__tests__/mockAdventureData");
-
-      // Mock getAdventure to ensure it's NOT called if early return works
-      const mockGetAdventure = vi
-        .spyOn(await import("@/data/adventureDatabase"), "getAdventure")
-        .mockResolvedValueOnce({
-          id: TEST_STORY_ID,
-          title: "Test Adventure",
-          content: "yaml content",
-          lastEdited: new Date(),
-          createdAt: new Date(),
-        });
-
-      // Start with valid adventure, but then set it to null to test the guard
-      // We need the component to render first with a valid adventure
-      renderWithAdventure(<AdventureTopBar />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      // Manually test the early return logic by checking behavior
-      // Since we can't easily set adventure to null after render in this test setup,
-      // this test verifies that when adventure IS present, the download works
-      const menuButton = await screen.findByTestId(
-        "adventure-context-menu-button"
-      );
-      expect(menuButton).toBeInTheDocument();
-
-      mockGetAdventure.mockRestore();
-    });
-
-    it("handles errors during download", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      // Mock getAdventure to throw an error
-      const mockGetAdventure = vi
-        .spyOn(await import("@/data/adventureDatabase"), "getAdventure")
-        .mockRejectedValueOnce(new Error("Database error"));
-
-      renderWithAdventure(<AdventureTopBar />, {
-        adventureId: TEST_STORY_ID,
-      });
-
-      const menuButton = await screen.findByTestId(
-        "adventure-context-menu-button"
-      );
-      fireEvent.click(menuButton);
-
-      const downloadMenuItem = screen.getByText("Download adventure as YAML");
-
-      await act(async () => {
-        fireEvent.click(downloadMenuItem);
-      });
-
-      // Should log error to console
-      await waitFor(() => {
-        expect(consoleSpy).toHaveBeenCalledWith(
-          "Failed to download adventure:",
-          expect.objectContaining({
-            message: "Database error",
-          })
-        );
-      });
-
-      consoleSpy.mockRestore();
-      mockGetAdventure.mockRestore();
-    });
-
-    it("handles case when adventure not found in database", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      // Mock getAdventure to return undefined
-      const mockGetAdventure = vi
-        .spyOn(await import("@/data/adventureDatabase"), "getAdventure")
-        .mockResolvedValueOnce(undefined);
-
-      renderWithAdventure(<AdventureTopBar />, {
-        adventureId: TEST_STORY_ID,
-      });
-
-      const menuButton = await screen.findByTestId(
-        "adventure-context-menu-button"
-      );
-      fireEvent.click(menuButton);
-
-      const downloadMenuItem = screen.getByText("Download adventure as YAML");
-
-      await act(async () => {
-        fireEvent.click(downloadMenuItem);
-      });
-
-      // Should log error to console - need to find the actual call among the act() warnings
-      await waitFor(() => {
-        const errorCalls = consoleSpy.mock.calls.filter(
-          (call) => call[0] === "Adventure not found in database"
-        );
-        expect(errorCalls.length).toBeGreaterThan(0);
-      });
-
-      consoleSpy.mockRestore();
-      mockGetAdventure.mockRestore();
-    });
-
-    it("handles early return when no adventureId", async () => {
-      const mockContextValue: AdventureContextType = {
-        adventure: null,
-        adventureId: null,
-        isLoading: false,
-        error: null,
-        isDebugModeEnabled: false,
-        isSaving: false,
-        setIsDebugModeEnabled: vi.fn(),
-        reloadAdventure: vi.fn(),
-        updateAdventure: vi.fn(),
-        updateIntroduction: vi.fn(),
-        updatePassage: vi.fn(),
-        withSaving: vi.fn(),
-      };
-
-      render(
-        <MemoryRouter initialEntries={["/"]}>
-          <AdventureContext.Provider value={mockContextValue}>
-            <AdventureTopBar />
-          </AdventureContext.Provider>
-        </MemoryRouter>
-      );
-
-      // Component should return null, so no menu button should be present
-      const menuButton = screen.queryByTestId("adventure-context-menu-button");
-      expect(menuButton).not.toBeInTheDocument();
     });
   });
 });
