--- src/components/TestAdventure/__tests__/TestAdventure.test.tsx
@@ -1,705 +1,0 @@
-import { screen, fireEvent, waitFor } from "@testing-library/react";
-import { vi, beforeEach, describe, it, expect } from "vitest";
-import { TestAdventure } from "../TestAdventure";
-import { renderWithAdventure } from "@/__tests__/testUtils";
-import { mockAdventure } from "@/__tests__/mockAdventureData";
-import {
-  getAdventureTestPassageRoute,
-  getAdventureTestRoute,
-} from "@/constants/routes";
-import { ErrorBoundary } from "@/components/ErrorBoundary/ErrorBoundary";
-import {
-  INTRODUCTION_TEST_IDS,
-  PASSAGE_TEST_IDS,
-  getChoiceButtonTestId,
-} from "../testIds";
-import type { Adventure } from "@/data/types";
-
-const TEST_STORY_ID = "test-adventure-id";
-
-// Mock react-router-dom navigate function
-const mockNavigate = vi.fn();
-let mockParams = { id: "1", adventureId: TEST_STORY_ID };
-
-vi.mock("react-router-dom", async () => {
-  const actual = await vi.importActual("react-router-dom");
-  return {
-    ...actual,
-    useNavigate: () => mockNavigate,
-    useParams: () => mockParams,
-  };
-});
-
-describe("TestAdventure Component", () => {
-  beforeEach(() => {
-    mockNavigate.mockClear();
-    vi.clearAllMocks();
-    mockParams = { id: "1", adventureId: TEST_STORY_ID };
-  });
-
-  describe("Introduction Mode", () => {
-    it("renders introduction when no passage id is provided", async () => {
-      mockParams = {
-        id: undefined as unknown as string,
-        adventureId: TEST_STORY_ID,
-      };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      const introContainer = await screen.findByTestId(
-        INTRODUCTION_TEST_IDS.CONTAINER
-      );
-      expect(introContainer).toBeInTheDocument();
-
-      const introTitle = await screen.findByTestId(INTRODUCTION_TEST_IDS.TITLE);
-      expect(introTitle).toBeInTheDocument();
-      expect(introTitle).toHaveTextContent(mockAdventure.metadata.title);
-
-      const introText = await screen.findByTestId(INTRODUCTION_TEST_IDS.TEXT);
-      expect(introText).toBeInTheDocument();
-    });
-
-    it("navigates to passage 1 when start button is clicked in introduction mode", async () => {
-      mockParams = {
-        id: undefined as unknown as string,
-        adventureId: TEST_STORY_ID,
-      };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      const button = await screen.findByTestId(
-        INTRODUCTION_TEST_IDS.START_BUTTON
-      );
-      fireEvent.click(button);
-
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestPassageRoute(TEST_STORY_ID, 1)
-      );
-    });
-
-    it("starts with empty inventory when introduction is displayed", async () => {
-      mockParams = {
-        id: undefined as unknown as string,
-        adventureId: TEST_STORY_ID,
-      };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-        isDebugModeEnabled: true,
-      });
-
-      // Verify introduction renders (inventory is in sidebar, managed by component state)
-      const introContainer = await screen.findByTestId(
-        INTRODUCTION_TEST_IDS.CONTAINER
-      );
-      expect(introContainer).toBeInTheDocument();
-
-      // In debug mode, verify no items are checked in the inventory
-      const item1Toggle = screen.queryByRole("switch", {
-        name: /Mock Item One/,
-      });
-      const item2Toggle = screen.queryByRole("switch", {
-        name: /Mock Item Two/,
-      });
-
-      if (item1Toggle) {
-        expect(item1Toggle).not.toBeChecked();
-      }
-      if (item2Toggle) {
-        expect(item2Toggle).not.toBeChecked();
-      }
-    });
-
-    it("navigates to correct passage when choice is clicked", async () => {
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      const firstChoice = await screen.findByTestId(getChoiceButtonTestId(0));
-      expect(firstChoice).toHaveAttribute("data-goto", "2");
-
-      fireEvent.click(firstChoice);
-
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestPassageRoute(TEST_STORY_ID, 2)
-      );
-    });
-  });
-
-  describe("Ending Passages", () => {
-    it("navigates to introduction when restart is clicked", async () => {
-      mockParams = { id: "4", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      const restartButton = await screen.findByTestId(
-        PASSAGE_TEST_IDS.RESTART_BUTTON
-      );
-      fireEvent.click(restartButton);
-
-      // Component navigates to introduction, which will reset state on remount
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestRoute(TEST_STORY_ID)
-      );
-    });
-  });
-
-  describe("Effects Handling", () => {
-    it("adds item to inventory when passage has add_item effect", async () => {
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          7: {
-            paragraphs: ["You found a magical sword!"],
-            choices: [{ text: "Continue", goto: 1 }],
-            effects: [{ type: "add_item", item: "sword" }],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "sword", name: "Magical Sword" },
-        ],
-      };
-
-      mockParams = { id: "7", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      // Wait for passage to render
-      await screen.findByText("You found a magical sword!");
-
-      // In debug mode, check that the item appears in inventory
-      await waitFor(() => {
-        const itemToggle = screen.queryByRole("switch", {
-          name: /Magical Sword/,
-        });
-        if (itemToggle) {
-          expect(itemToggle).toBeChecked();
-        }
-      });
-    });
-
-    it("removes item from inventory when passage has remove_item effect", async () => {
-      // Create an adventure where passage 8 removes an item
-      // We won't actually test the removal works across navigation,
-      // just that the component handles the effect without errors
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          8: {
-            paragraphs: ["You used the healing potion."],
-            choices: [{ text: "Continue", goto: 1 }],
-            effects: [{ type: "remove_item", item: "potion" }],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "potion", name: "Healing Potion" },
-        ],
-      };
-
-      mockParams = { id: "8", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByText("You used the healing potion.");
-
-      // Just verify the passage renders without errors
-      // The effect is applied via onRemoveItem callback
-      expect(
-        screen.getByTestId(PASSAGE_TEST_IDS.CONTAINER)
-      ).toBeInTheDocument();
-    });
-
-    it("handles multiple effects in a passage", async () => {
-      // Test that a passage with multiple effects renders without errors
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          9: {
-            paragraphs: ["You traded the gem for a key."],
-            choices: [{ text: "Continue", goto: 1 }],
-            effects: [
-              { type: "remove_item", item: "gem" },
-              { type: "add_item", item: "key" },
-            ],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "gem", name: "Ruby Gem" },
-          { id: "key", name: "Golden Key" },
-        ],
-      };
-
-      mockParams = { id: "9", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByText("You traded the gem for a key.");
-
-      // Verify passage renders and the key was added
-      await waitFor(() => {
-        const keyToggle = screen.queryByRole("switch", { name: /Golden Key/ });
-        if (keyToggle) {
-          expect(keyToggle).toBeChecked();
-        }
-      });
-    });
-
-    it("does not apply effects for ending passages", async () => {
-      // Test that ending passages don't trigger effects by checking a proper ending passage
-      // Passage 4 in mockAdventure is an ending (it has ending: true, no choices, no effects)
-      mockParams = { id: "4", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-        isDebugModeEnabled: true,
-      });
-
-      // Wait for the ending passage to render
-      await screen.findByTestId(PASSAGE_TEST_IDS.RESTART_BUTTON);
-
-      // Verify none of the test items were added (ending passages don't process effects)
-      const item1Toggle = screen.queryByRole("switch", {
-        name: /Mock Item One/,
-      });
-      const item2Toggle = screen.queryByRole("switch", {
-        name: /Mock Item Two/,
-      });
-
-      if (item1Toggle) {
-        expect(item1Toggle).not.toBeChecked();
-      }
-      if (item2Toggle) {
-        expect(item2Toggle).not.toBeChecked();
-      }
-    });
-
-    it("handles passages without effects", async () => {
-      // Use existing passage 1 which has no effects
-      mockParams = { id: "1", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByTestId(PASSAGE_TEST_IDS.CONTAINER);
-
-      // Inventory should remain unchanged (empty or with existing items)
-      // This test just verifies the component renders without errors
-      expect(
-        screen.getByTestId(PASSAGE_TEST_IDS.CONTAINER)
-      ).toBeInTheDocument();
-    });
-
-    it("does not execute effects when passageId is invalid", async () => {
-      mockParams = { id: "not-a-number", adventureId: TEST_STORY_ID };
-
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          adventure: mockAdventure,
-        }
-      );
-
-      // Should show error instead of executing effects
-      const errorMessages = await screen.findAllByText(/is not valid/);
-      expect(errorMessages.length).toBeGreaterThan(0);
-
-      consoleSpy.mockRestore();
-    });
-  });
-
-  describe("Error Handling", () => {
-    it("renders introduction view when id param is undefined", async () => {
-      mockParams = {
-        id: undefined as unknown as string,
-        adventureId: TEST_STORY_ID,
-      };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-      });
-
-      // Should render introduction instead of throwing an error
-      const introContainer = await screen.findByTestId(
-        INTRODUCTION_TEST_IDS.CONTAINER
-      );
-      expect(introContainer).toBeInTheDocument();
-
-      const introTitle = await screen.findByTestId(INTRODUCTION_TEST_IDS.TITLE);
-      expect(introTitle).toBeInTheDocument();
-      expect(introTitle).toHaveTextContent(mockAdventure.metadata.title);
-
-      const startButton = await screen.findByTestId(
-        INTRODUCTION_TEST_IDS.START_BUTTON
-      );
-      expect(startButton).toBeInTheDocument();
-    });
-
-    it("throws AdventureLoadError when there is a load error", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          error: "Failed to load adventure",
-        }
-      );
-
-      expect(
-        (await screen.findAllByText("Failed to load adventure")).length
-      ).toBeGreaterThan(0);
-      expect(screen.getByText("A system error occurred")).toBeInTheDocument();
-
-      consoleSpy.mockRestore();
-    });
-
-    it("throws AdventureNotFoundError when adventure is not found", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          adventure: null,
-        }
-      );
-
-      expect(
-        (await screen.findAllByText("Adventure not found.")).length
-      ).toBeGreaterThan(0);
-      expect(screen.getByText("A system error occurred")).toBeInTheDocument();
-
-      consoleSpy.mockRestore();
-    });
-
-    it("throws InvalidPassageIdError for non-numeric passage IDs", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      mockParams = { id: "invalid", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          adventure: mockAdventure,
-        }
-      );
-
-      const errorMessages = await screen.findAllByText(/is not valid/);
-      expect(errorMessages.length).toBeGreaterThan(0);
-      expect(screen.getByText("A system error occurred")).toBeInTheDocument();
-
-      consoleSpy.mockRestore();
-    });
-
-    it("throws InvalidPassageIdError for negative passage IDs", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      mockParams = { id: "-5", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          adventure: mockAdventure,
-        }
-      );
-
-      const errorMessages = await screen.findAllByText(/is not valid/);
-      expect(errorMessages.length).toBeGreaterThan(0);
-      expect(screen.getByText("A system error occurred")).toBeInTheDocument();
-
-      consoleSpy.mockRestore();
-    });
-
-    it("throws PassageNotFoundError when passage does not exist", async () => {
-      const consoleSpy = vi
-        .spyOn(console, "error")
-        .mockImplementation(() => {});
-
-      mockParams = { id: "999", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(
-        <ErrorBoundary>
-          <TestAdventure />
-        </ErrorBoundary>,
-        {
-          adventureId: TEST_STORY_ID,
-          adventure: mockAdventure,
-        }
-      );
-
-      const errorMessages = await screen.findAllByText(/does not exist/);
-      expect(errorMessages.length).toBeGreaterThan(0);
-      expect(screen.getByText("A system error occurred")).toBeInTheDocument();
-
-      consoleSpy.mockRestore();
-    });
-  });
-
-  describe("Debug Mode Choice Display", () => {
-    it("maintains correct navigation when debug mode is enabled", async () => {
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: mockAdventure,
-        isDebugModeEnabled: true,
-      });
-
-      const firstChoice = await screen.findByTestId(getChoiceButtonTestId(0));
-      expect(firstChoice).toHaveTextContent("2: Go to mock passage 2");
-      expect(firstChoice).toHaveAttribute("data-goto", "2");
-
-      fireEvent.click(firstChoice);
-
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestPassageRoute(TEST_STORY_ID, 2)
-      );
-    });
-  });
-
-  describe("Inventory State Management", () => {
-    it("resets inventory when navigating from passage back to introduction", async () => {
-      // This test verifies that clicking the restart button on an ending passage
-      // navigates to the introduction (where inventory state is reset)
-      const adventureWithEnding: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          10: {
-            paragraphs: ["The adventure concludes!", "The end!"],
-            ending: true,
-          },
-        },
-      };
-
-      mockParams = { id: "10", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEnding,
-        isDebugModeEnabled: true,
-      });
-
-      // Wait for ending passage to render
-      await screen.findByText("The adventure concludes!");
-
-      // Click the restart button to navigate to introduction
-      const restartButton = screen.getByTestId(PASSAGE_TEST_IDS.RESTART_BUTTON);
-      fireEvent.click(restartButton);
-
-      // Verify navigate was called to go to introduction (where inventory resets)
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestRoute(TEST_STORY_ID)
-      );
-    });
-
-    it("properly handles remove_item effect during passage navigation", async () => {
-      // Create adventure where we add an item in one passage and remove it in another
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          11: {
-            paragraphs: ["You picked up a torch."],
-            choices: [{ text: "Continue", goto: 12 }],
-            effects: [{ type: "add_item", item: "torch" }],
-          },
-          12: {
-            paragraphs: ["The torch burned out and crumbled to ash."],
-            choices: [{ text: "Move on", goto: 1 }],
-            effects: [{ type: "remove_item", item: "torch" }],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "torch", name: "Burning Torch" },
-        ],
-      };
-
-      // Start at passage 11 to add the item
-      mockParams = { id: "11", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByText("You picked up a torch.");
-
-      // Verify item was added
-      await waitFor(() => {
-        const torchToggle = screen.queryByRole("switch", {
-          name: /Burning Torch/,
-        });
-        if (torchToggle) {
-          expect(torchToggle).toBeChecked();
-        }
-      });
-
-      // Directly navigate to passage 12 by clicking the choice
-      const continueButton = await screen.findByRole("button", {
-        name: /Continue/,
-      });
-      fireEvent.click(continueButton);
-
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestPassageRoute(TEST_STORY_ID, 12)
-      );
-    });
-
-    it("does not duplicate items when add_item effect adds same item twice", async () => {
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          13: {
-            paragraphs: ["You tried to pick up the coin again."],
-            choices: [{ text: "Continue", goto: 1 }],
-            effects: [
-              { type: "add_item", item: "coin" },
-              { type: "add_item", item: "coin" }, // Try to add same item twice
-            ],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "coin", name: "Gold Coin" },
-        ],
-      };
-
-      mockParams = { id: "13", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByText("You tried to pick up the coin again.");
-
-      // Verify the item is in inventory (should only appear once)
-      await waitFor(() => {
-        const coinToggle = screen.queryByRole("switch", {
-          name: /Gold Coin/,
-        });
-        if (coinToggle) {
-          expect(coinToggle).toBeChecked();
-        }
-      });
-    });
-
-    it("executes remove_item effect correctly when item exists in inventory", async () => {
-      // This test ensures the remove_item branch is executed properly
-      const adventureWithEffects: Adventure = {
-        ...mockAdventure,
-        passages: {
-          ...mockAdventure.passages,
-          14: {
-            paragraphs: ["You found a mysterious key."],
-            choices: [{ text: "Use the key", goto: 15 }],
-            effects: [{ type: "add_item", item: "key" }],
-          },
-          15: {
-            paragraphs: ["You used the key to open the door. It broke!"],
-            choices: [{ text: "Enter", goto: 1 }],
-            effects: [{ type: "remove_item", item: "key" }],
-          },
-        },
-        items: [
-          ...(mockAdventure.items || []),
-          { id: "key", name: "Mysterious Key" },
-        ],
-      };
-
-      // Start at passage 14 where we get the key
-      mockParams = { id: "14", adventureId: TEST_STORY_ID };
-
-      renderWithAdventure(<TestAdventure />, {
-        adventureId: TEST_STORY_ID,
-        adventure: adventureWithEffects,
-        isDebugModeEnabled: true,
-      });
-
-      await screen.findByText("You found a mysterious key.");
-
-      // Wait for the add_item effect to execute
-      await waitFor(() => {
-        const keyToggle = screen.queryByRole("switch", {
-          name: /Mysterious Key/,
-        });
-        if (keyToggle) {
-          expect(keyToggle).toBeChecked();
-        }
-      });
-
-      // Now click to navigate to passage 15 where the key is removed
-      const useKeyButton = await screen.findByRole("button", {
-        name: /Use the key/,
-      });
-      fireEvent.click(useKeyButton);
-
-      // This should trigger the remove_item effect
-      expect(mockNavigate).toHaveBeenCalledWith(
-        getAdventureTestPassageRoute(TEST_STORY_ID, 15)
-      );
-    });
-  });
-});
